HW4

+++ src/main/java/com/example/S2_H1/service/UserService.java
@@ -12,6 +12,9 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
 @Slf4j
 @Service
 @AllArgsConstructor
@@ -21,11 +24,15 @@ public class UserService {
   private final CategoryRepository categoryRepository;
   private final SiteRepository siteRepository;
 
+  private final Map<Long, String> userNames = new ConcurrentHashMap<>();
+
   public UserId registerUser(UserDto userDto) {
     log.info("Создание нового пользователя");
     User user = User.builder().userName(userDto.getUserName()).password(userDto.getUserPassword()).build();
     log.info("Пользователь создан");
-    return userRepository.saveAccount(user);
+    UserId userId = userRepository.saveAccount(user);
+    userNames.put(userId.id(), user.getUserName());
+    return userId;
   }
 
   public void deleteUser(Long parsUserId) {
@@ -33,25 +40,32 @@ public class UserService {
     UserId userId = new UserId(parsUserId);
     try {
       userRepository.deleteAccount(userId);
-    log.info("Удаление категорий пользователя с айди {}", parsUserId);
-    categoryRepository.deleteByUserId(userId);
-    log.info("Удаление сайтов пользователя с айди {}", parsUserId);
-    siteRepository.deleteByUserId(userId);
+      userNames.remove(userId.id());
+      log.info("Удаление категорий пользователя с айди {}", parsUserId);
+      categoryRepository.deleteByUserId(userId);
+      log.info("Удаление сайтов пользователя с айди {}", parsUserId);
+      siteRepository.deleteByUserId(userId);
     } catch (Exception e) {
       log.info("Пользователь с айди {} не был найден", parsUserId);
       throw new NoSuchUserException("Данный пользователь не найден", e);
     }
   }
 
+  //Exactly Once
   public User updateUserName(UserUpdateNameDto userUpdateNameDto, Long parsUserId) {
     String newUserName = userUpdateNameDto.getNewUserName();
     UserId userId = new UserId(parsUserId);
+    if (userNames.get(parsUserId).equals(newUserName)) {
+      log.info("Новое имя пользователя совпадает с текущим");
+      return userRepository.getUser(userId);
+    }
     log.info("Обновление имени юзера с айди {}, на {}", parsUserId, newUserName);
     User userWithOldName = userRepository.getUser(userId);
     User userWithUpdatedName = User.builder().userId(userId).password(userWithOldName.getPassword()).userName(newUserName).build();
     log.info("Имя пользователя обновлено");
     userRepository.deleteAccount(userId);
     userRepository.saveUserWithoutIdUpdate(userWithUpdatedName);
+    userNames.put(parsUserId, userWithUpdatedName.getUserName());
     log.info("Изменения сохранены в репозиторий");
     return userWithUpdatedName;
   }
@@ -63,6 +77,7 @@ public class UserService {
     User userWithUpdatedData = User.builder().userId(userId).userName(userDto.getUserName()).password(userDto.getUserPassword()).build();
     log.info("Данные пользователя обновлены");
     userRepository.saveUserWithoutIdUpdate(userWithUpdatedData);
+    userNames.put(parsUserId, userWithUpdatedData.getUserName());
     log.info("Изменения сохранены в репозиторий");
     return userWithUpdatedData;
   }
